Select & Main Edits
In order to use this, you'll probably want to reference lines 11-21 which are commented out.

Essentially, you'll need to specify the destination of the from "temp2" to whatever database you are wishing to connect to.

The edits to the 'select()' method, were to just test the select method. Currently I don't have a table set up with the columns that are specified in the select method, so I commented out the parts that would cause errors.