Methods created (read for info)
Ok so, a lot of changes. These changes make it to where we can get the history with getMessages. Currently this will print out the entire history of a chatroom given the chat name.

Eventually a similar method will be made for retrieving strictly new messages. This getMessages() method increments your known messages which will aid in the retrieval of new messages with a new method later down the road.

Working next on returning the user as well with the message.