Holy shid this hot garbage! (but I can't stop  listening)
Ok so there is a lot of changes that were made to database which were made to print the new messages, there are some comments along the way.

Basic gist - make a query to get all the "id's" associated with the messages for a chatroom, add those ids to a list. Print out the index of the current knownMessages and beyond of that arraylist.

Also need to add total known message increaser to this