Update Main.java
This is a fix for the error when you send just hit enter or attempt to spam with spaces. This will simply not allow messages of only blank space

Print newMessages is called first, and then the new message that was typed in the input is inserted to the database. This causes a behavior of, when there is a new message from a new user, the other user will:
1) Send a new message
2) Before the message is insertted, the other message from the other user is printed
3) then the user's message is insertted.