Error when printing history upon joining a chatroom with no messages
This would happen when the user created a chatroom, and for one reason or another, left the chatroom before sending any messages. This would cause an error because we cannot convert the first message Id (null) to an int. This one line checks if there are messages in the chat - if not, return.